First try at static variables in clojure
Static variables aren't really a thing in clojure. There's a static
keyword which doesn't really mean the same thing as it does in Java.
I think the closest you get is just by def' a var in side a namespace:
the value is there and you don't have to instantiate a class.

I may spruce this up a little more to match exactly whats the java verion
of this is.